Move dailyPageInfor from Service to Controler
